tested adn runig
