+++ testinggit_API/urls.py
@@ -3,5 +3,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('api.urls')),
+    path('accounts/', include('accounts.urls')),
 ]
